% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dlm.R
\name{create.fitted.model}
\alias{create.fitted.model}
\title{create.fitted.model}
\usage{
create.fitted.model(data, trend.order = NULL, seasonal.periods = NULL)
}
\arguments{
\item{data}{the data fit the model with}

\item{trend.order}{the order of the trend (set to NULL or 0 for no trend)}

\item{seasonal.periods}{the period of the seasonal components as a list}
}
\value{
returns a "fitted.dlm" object
}
\description{
fit a DLM model on data with trend and seasonal components
}
