% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dlm.R
\name{create.fitted.st.model}
\alias{create.fitted.st.model}
\title{create.fitted.st.model}
\usage{
create.fitted.st.model(data, seasonal.period)
}
\arguments{
\item{data}{the data fit the model with}

\item{seasonal.period}{the period of the seasonal component}
}
\value{
returns a "fitted.dlm" object
}
\description{
fit a DLM model on data with 1st order trend and a single seasonal component
}
