% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dlm.R
\name{dlm.build.trend.seas}
\alias{dlm.build.trend.seas}
\title{dlm.build.trend.seas}
\usage{
dlm.build.trend.seas(parm, seasonal.period)
}
\arguments{
\item{parm}{a list of three parameters that parameterize the dlm}

\item{seasonal.period}{the period of the seasonal component}
}
\value{
returns a dlm object
}
\description{
build a DLM with a trend and a seasonal component
}
