% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dlm.R
\name{dlm.fit.prealloc}
\alias{dlm.fit.prealloc}
\title{dlm.fit.prealloc}
\usage{
dlm.fit.prealloc(data, seasonal.period, hessian = TRUE)
}
\arguments{
\item{data}{the data to fit the model on}

\item{seasona.period}{the period of the seasonal component}
}
\value{
returns the MLE fit for the DLM
}
\description{
fit a DLM with a trend and seasonal component while preallocating memory for the DLM matricies
}
