% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dlm.R
\name{dlm.fit.trend.seas}
\alias{dlm.fit.trend.seas}
\title{dlm.fit.trend.seas}
\usage{
dlm.fit.trend.seas(data, seasonal.period, hessian = TRUE)
}
\arguments{
\item{data}{the data to fit the model on}

\item{seasona.period}{the period of the seasonal component}
}
\value{
returns the MLE fit for the DLM
}
\description{
fit a DLM with a trend and seasonal component
}
