% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dlm.R
\name{dlm.get.build.trend.seas}
\alias{dlm.get.build.trend.seas}
\title{dlm.get.build.trend.seas}
\usage{
dlm.get.build.trend.seas(seasonal.period)
}
\arguments{
\item{seasonal.period}{the period of the seasonal component}

\item{data}{the data to fit the model with}
}
\value{
returns a build function that takes in a list of 3 parameters as a list and returns a dlm
}
\description{
returns a build function that builds a DLM with a trend and a seasonal component
The build function takes in a vector of parameters parm
}
