% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_results.R
\name{get.bats.results}
\alias{get.bats.results}
\title{get.bats.results}
\usage{
get.bats.results(fit, nAhead = NULL)
}
\arguments{
\item{fit}{the fitted bats model}

\item{nAhead}{how many values ahead to forecast}
}
\value{
results as a named list with fields $fitted.values, $residuals, and optionally $forecast
}
\description{
produces fitted values, predicted values, etc for the BATS model
}
