% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_results.R
\name{get.dlm.results}
\alias{get.dlm.results}
\title{get.dlm.results}
\usage{
get.dlm.results(fit, data, nAhead = NULL)
}
\arguments{
\item{fit}{the fitted dlm model}

\item{data}{the training data for the dlm}

\item{nAhead}{how many values ahead to forecast}
}
\value{
results as a named list with fields $fitted.values, $residuals, and optionally $forecast
}
\description{
produces fitted values, predicted values, etc for a DLM model
}
