% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_results.R
\name{get.mae}
\alias{get.mae}
\title{get.mae}
\usage{
get.mae(errors)
}
\arguments{
\item{errors}{the error values}
}
\value{
the MAE
}
\description{
take in a list of errors and return the mean absolute error
}
