% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_results.R
\name{get.mse}
\alias{get.mse}
\title{get.mse}
\usage{
get.mse(errors)
}
\arguments{
\item{errors}{the error values}
}
\value{
the MSE
}
\description{
take in a list of errors and return the mean squared error
}
