% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_plots.R
\name{test.comparison.plots}
\alias{test.comparison.plots}
\alias{test.comparision.plots}
\title{test.comparison.plots}
\usage{
test.comparision.plots(data.test, forecasted.values, model.name)
}
\arguments{
\item{data.test}{test data}

\item{forecasted.values}{forecasted values}

\item{model.name}{the name of the model you trained}
}
\value{
NONE
}
\description{
show plots for comparing test data to forecast values
}
