% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_plots.R
\name{training.comparison.plots}
\alias{training.comparison.plots}
\title{training.comparison.plots}
\usage{
training.comparison.plots(data.train, fitted.values, residuals, model.name)
}
\arguments{
\item{data.train}{training data}

\item{fitted.values}{fitted values}

\item{residuals}{residuals}

\item{model.name}{the name of the model you trained}
}
\value{
NONE
}
\description{
show plots for comparing training data to fitted values and residuals
}
